requirements.txt
django
django-crispy-forms
crispy-bootstrap4
pillow
django-recaptcha
requests
psycopg2
djangorestframework
django-admin-confirm
django-axes
